Utilities_Format/document-summarizer/scripts/requirements.txt
pypdf
python-docx
python-pptx
openpyxl
tqdm>=4.65.0
pyyaml>=6.0